pyproject.toml
[project]
name = "openenv-block-blast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openenv-core>=0.2.3",
]
